Refactor transaction import parsing to avoid eager DB load
